refactor: fold utils/ into lib/ and root hooks/ into app/hooks/
Two three-file directories doing the same job as existing homes.
Pure move + import-path rewrite; no behavior change.
