+++ frontend/components/client/modals/githubfeedbackmodal.tsx
@@ -28,7 +28,7 @@ import {
 } from '@mui/joy';
 import { ChangeEvent, useState } from 'react';
 import { AccessibilityNew, BugReport, Build, Error as ErrorIcon, Event, GitHub, Info, Person } from '@mui/icons-material';
-import { useFormSubmission } from '@/hooks/useAsyncOperation';
+import { useFormSubmission } from '@/app/hooks/useAsyncOperation';
 import { useOrgCensusContext, usePlotContext, useSiteContext } from '@/app/contexts/compat-hooks';
 import { usePathname } from 'next/navigation';
 import { useSession } from 'next-auth/react';
