+++ frontend/config/macros/coreapifunctions.ts
@@ -5,7 +5,7 @@ import MapperFactory from '@/config/datamapper';
 import { handleUpsert } from '@/config/utils';
 import { format } from 'mysql2/promise';
 import { HTTPResponses } from '@/config/macros';
-import { handleError } from '@/utils/errorhandler';
+import { handleError } from '@/lib/errorhandler';
 import { FamilyResult, GenusResult, SpeciesResult } from '@/lib/db/definitions/taxonomies';
 import { getCookie } from '@/app/actions/cookiemanager';
 import { insertIngestionFailureRows } from '@/config/measurementerrors';
